requirements.txt
Flask==2.2.5
pandas==1.5.3
requests==2.31.0
groq==<version>
python-dotenv==1.0.0
plotly==5.15.0
